created a Human class and called in main file
