+++ connect_four/connect_four.rb
@@ -2,7 +2,7 @@ require './game'
 require './board'
 require './player'
 require './computer'
-# require './connect_four/human'
+require './human'
 
 
 # Includes the
