+++ connect_four/game.rb
@@ -14,11 +14,11 @@ class Game
 
     answer = gets.chomp.strip
     if(answer == "1")
-      @player1 = Player.new("Player1", :r, @board)
+      @player1 = Human.new("Player1", :r, @board)
       @player2 = Computer.new("Computer", :y, @board)
     elsif(answer == "2")
-      @player1 = Player.new("Player1", :r, @board)
-      @player2 = Player.new("Player2", :y, @board)
+      @player1 = Human.new("Player1", :r, @board)
+      @player2 = Human.new("Player2", :y, @board)
     else 
       puts "Please ensure you enter your input correctly"
       initial_setup
